@@ -4,12 +4,12 @@
   "private": true,
   "homepage": "https://lwatson2.github.io/react-pokedex",
   "dependencies": {
-    "@material-ui/core": "^3.2.0",
-    "@material-ui/icons": "^3.0.1",
-    "axios": "^0.18.0",
+    "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
+    "axios": "^0.18.1",
     "material-jsx-icons": "^1.0.0",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4"
@@ -24,6 +24,6 @@
   },
   "devDependencies": {
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "gh-pages": "^2.0.1"
+    "gh-pages": "^2.1.1"
   }
 }
